refactor: move VolunteerTopNav code to VolunteerManagement
